+++ src/App/MainMenu.swift
@@ -78,6 +78,11 @@ func setupMainMenu(target: AppDelegate) {
     archiveItem.image = NSImage(systemSymbolName: "archivebox", accessibilityDescription: nil)
     fileMenu.addItem(archiveItem)
 
+    let extractItem = NSMenuItem(title: "Extract Here", action: #selector(FileListViewController.extractArchive(_:)), keyEquivalent: "E")
+    extractItem.keyEquivalentModifierMask = [.command, .shift]
+    extractItem.image = NSImage(systemSymbolName: "arrow.up.bin", accessibilityDescription: nil)
+    fileMenu.addItem(extractItem)
+
     fileMenu.addItem(NSMenuItem.separator())
 
     let getInfoItem = NSMenuItem(title: "Get Info", action: #selector(FileListViewController.getInfo(_:)), keyEquivalent: "i")